Step Two implementation
Developed the Step One and Step Two as provided with extensive cleaning of the UCI Machine Learning Datasets(Casts, Movies, Award Types, Actors). 
Step Two is at the bottom of the "Scikit-Test.ipynb" that basically generates the predictions (Ratings of the movies) to the input data into a .csv file.
These predictions can then be made using the function with a movie name as a parameter that gives the output of the "Rating" as a result.